创建pod

+++ api/v1/springboot_types.go
@@ -31,7 +31,7 @@ type SpringBootSpec struct {
 	// Foo is an example field of SpringBoot. Edit springboot_types.go to remove/update
 	Foo   string `json:"foo,omitempty"`
 	Image string `json:"image,omitempty"`
-	Port  int16  `json:"port,omitempty"`
+	Port  int32  `json:"port,omitempty"`
 }
 
 // SpringBootStatus defines the observed state of SpringBoot
